Tests are extracted at separate directory
* tests/ introduction
* unittest discovery running at travis config
